Externalize DataModel Pipeline Interface
