+++ demo/index.html
@@ -8,6 +8,7 @@
     <script src="data/widedata.js"></script>
     <script src="build/fin-hypergrid.js"></script>
     <script src="js/demo/build/index.js"></script>
+    <script src="build/add-ons/datasource-pipelines.js"></script>
     <link rel="stylesheet" href="css/demo.css">
 
 </head>
