+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fin-hypergrid",
-  "version": "1.3.3",
+  "version": "1.4.0",
   "description": "Canvas-based high-performance spreadsheet",
   "repository": {
     "type": "git",
