+++ src/Hypergrid.js
@@ -29,7 +29,6 @@ var EDGE_STYLES = ['top', 'bottom', 'left', 'right'],
  * @param {string|Element} [container] - CSS selector or Element
  * @param {object} [options]
  * @param {function} [options.Behavior=behaviors.JSON] - A behavior constructor or instance
- * @param {function[]} [options.pipeline] - A list function constructors to use for passing data through a series of transforms to occur on reindex call
  * @param {function|object[]} [options.data] - Passed to behavior constructor. May be:
  * * An array of congruent raw data objects
  * * A function returning same
@@ -301,10 +300,6 @@ var Hypergrid = Base.extend('Hypergrid', {
      * @param {object} [options]
      * @param {object} [options.subgrids] - Consumed by {@link Behavior#reset}.
      * If omitted, previously established subgrids list is reused.
-     * @param {object} [options.pipeline] - Consumed by {@link dataModels.JSON#reset}.
-     * If omitted, previously established pipeline is reused.
-     * @param {object} [options.controllers] - Consumed by {@link dataModels.JSON#reset}.
-     * If omitted, previously established controllers list is reused.
      * @memberOf Hypergrid#
      */
     reset: function(options) {
@@ -336,8 +331,6 @@ var Hypergrid = Base.extend('Hypergrid', {
         options = options || {};
         this.behavior.reset({
             subgrids: options.subgrids,
-            pipeline: options.pipeline,
-            controllers: options.controllers
         });
 
         this.renderer.reset();
@@ -723,7 +716,6 @@ var Hypergrid = Base.extend('Hypergrid', {
      * @param {object} options - _(See {@link behaviors.JSON#setData}.)_
      * @param {Behavior} [options.behavior=behaviors.JSON] - The behavior (model) can be either a constructor or an instance.
      * @param {dataRowObject[]} [options.data] - _(See {@link behaviors.JSON#setData}.)_
-     * @param {pipelineSchema} [options.pipeline] - New pipeline description.
      */
     setBehavior: function(options) {
         if (!this.behavior) {
@@ -735,7 +727,6 @@ var Hypergrid = Base.extend('Hypergrid', {
             this.initCanvas();
             this.initScrollbars();
             this.refreshProperties();
-            this.behavior.reindex();
         }
     },
 
@@ -778,11 +769,10 @@ var Hypergrid = Base.extend('Hypergrid', {
 
     /**
      * @memberOf Hypergrid#
-     * @param {object} [pipelines] - New pipeline description. _(See {@link dataModels.JSON#setPipeline}.)_
-     * @param {object} [options] - _(See {@link dataModels.JSON#setPipeline}.)_
+     * @deprecated
      */
     setPipeline: function(DataSources, options){
-        this.behavior.setPipeline(DataSources, options);
+        return this.deprecated('setPipeline()', '', '1.4.0', arguments, 'No Longer Supported');
     },
 
     /**
@@ -1641,7 +1631,7 @@ var Hypergrid = Base.extend('Hypergrid', {
     },
 
     getUnfilteredRowCount: function() {
-        return this.deprecated('getUnfilteredRowCount()', null, '1.2.0', arguments, 'No longer supported');
+        return this.deprecated('getUnfilteredRowCount()', '', '1.2.0', arguments, 'No longer supported');
     },
 
     /**
@@ -2068,53 +2058,33 @@ var Hypergrid = Base.extend('Hypergrid', {
     },
 
     /**
-     * @summary Get the given data controller.
-     * @param {string} type
-     * @returns {null|undefined|*} The data controller or:
-     * * `null` means unknown data controller.
-     * * `undefined` means the data source handles this data controller but the data controller is undefined.
+     * @deprecated
      * @memberOf Hypergrid#
      */
     getController: function(type) {
-        return this.behavior.getController(type);
+        return this.deprecated('getController()', '', 'v1.4.0', arguments, 'No Longer Supported');
     },
 
     /**
-     * @summary Set the given data controller(s).
-     * @desc The data model needs to be able to accept the specified data controller type(s). If it fails to accept the specified data controller(s), an error condition is raised.
-     *
-     * (To ignore the error, place the call in a `try...catch`. From there you could call {@link Base#notify|notify} to report it as a warning or an alert instead.)
-     *
-     * Setting data controller(s) triggers a shape change.
-     * @param {string} typeOrHashOfTypes - One of:
-     * * **object** - Hash of multiple data controllers, by type.
-     * * **string** - Type of the single data controller given in `controller`.
-     * @param {dataControlInterface} [controller] - Only required when 'hash' is a string; omit when `hash` is an object.
-     * @returns {object} - Hash of all results, by type. Each member will be:
-     * * The given data controller for that type when defined.
-     * * A new "null" data controller, generated by the data model when the given data controller for that type was `undefined`.
-     * * `undefined` - The data controller was unknown to the data model.
+     * @deprecated
      * @memberOf Hypergrid#
      */
     setController: function(typeOrHashOfTypes, controller) {
-        var results = this.behavior.setController(typeOrHashOfTypes, controller),
-            rejections = Object.keys(results).filter(function(name) {
-                return !results[name];
-            }).toString();
-
-        if (rejections) {
-            throw 'Unexpected data controller(s): ' + rejections;
-        }
-
-        return results;
+        return this.deprecated('setController()', '', 'v1.4.0', arguments, 'No Longer Supported');
     },
-
+    /**
+     * @deprecated
+     * @memberOf Hypergrid#
+     */
     prop: function(type, columnIndex, keyOrHash, value) {
-        return this.behavior.prop.apply(this.behavior, arguments);
+        return this.deprecated('prop()', '', 'v1.4.0', arguments, 'No Longer Supported');
     },
-
+    /**
+     * @deprecated
+     * @memberOf Hypergrid#
+     */
     get charMap() {
-        return this.behavior.charMap;
+        return this.deprecated('.charMap', '', 'v1.4.0', arguments, 'No Longer Supported');
     }
 });
