+++ src/behaviors/Behavior.js
@@ -86,19 +86,11 @@ var Behavior = Base.extend('Behavior', {
      * @param {object} [options]
      * @param {object} [options.pipeline] - Consumed by {@link dataModels.JSON#reset}.
      *
-     * If omitted, previously established pipeline is reused.
-     * @param {object} [options.controllers] - Consumed by {@link dataModels.JSON#reset}.
-     *
-     * If omitted, previously established controllers list is reused.
-     * @memberOf Behavior#
      */
     reset: function(options) {
         if (this.dataModel) {
             options = options || {};
-            this.dataModel.reset({
-                pipeline: options.pipeline,
-                controllers: options.controllers
-            });
+            this.dataModel.reset();
         } else {
             /**
              * @type {dataModelAPI}
@@ -253,11 +245,11 @@ var Behavior = Base.extend('Behavior', {
     },
 
     /**
+     * @deprecated
      * @memberOf Behavior#
      */
     reindex: function() {
-        this.dataModel.reindex();
-        this.shapeChanged();
+        return this.deprecated('reindex()', null, '1.4.0', arguments, 'No Longer Supported');
     },
 
     /**
@@ -285,6 +277,7 @@ var Behavior = Base.extend('Behavior', {
 
     /**
      * @memberOf Behavior#
+     * @deprecated
      * @desc getter for a [Memento](http://c2.com/cgi/wiki?MementoPattern) Object
      * @returns {object}
      */
@@ -349,6 +342,10 @@ var Behavior = Base.extend('Behavior', {
         }
     },
 
+    /**
+     * @memberOf Behavior#
+     * @deprecated
+     */
     _setColumnOrder: function(columnIndexes) {
         this.deprecated('_setColumnOrder(columnIndexes)', 'setColumnOrder(columnIndexes)', '1.2.10', arguments);
     },
@@ -519,10 +516,12 @@ var Behavior = Base.extend('Behavior', {
     getRowCount: function() {
         return this.dataModel.getRowCount();
     },
-
-    getUnfilteredValue: function(x, y) {
-        var column = this.getActiveColumn(x);
-        return column && column.getUnfilteredValue(y);
+    /**
+     * @memberOf Behavior#
+     * @deprecated
+     */
+    getUnfilteredValue: function() {
+        return this.deprecated(' getUnfilteredValue:()', null, '1.4.0', arguments, 'No Longer Supported');
     },
 
     /**
@@ -543,10 +542,17 @@ var Behavior = Base.extend('Behavior', {
         xOrCellEvent.value = value;
     },
 
+    /**
+     * @memberOf Behavior#
+     * @deprecated
+     */
     getDataValue: function(x, y) {
         return this.deprecated('getDataValue(x, y)', 'dataModel.getValue(x, y)', '1.1.0', arguments);
     },
-
+    /**
+     * @memberOf Behavior#
+     * @deprecated
+     */
     setDataValue: function(x, y, value) {
         return this.deprecated('setDataValue(x, y, value)', 'dataModel.setValue(x, y, value)', '1.1.0', arguments);
     },
@@ -716,6 +722,7 @@ var Behavior = Base.extend('Behavior', {
 
     /**
      * @memberOf Behavior#
+     * @deprecated
      * @desc The value is lazily initialized and comes from the properties mechanism for '`defaultRowHeight`', which should be ~20px.
      * @returns {number} The row height in pixels.
      */
@@ -1128,6 +1135,10 @@ var Behavior = Base.extend('Behavior', {
     getActiveColumnCount: function() {
         return this.columns.length;
     },
+    /**
+     * @memberOf Behavior#
+     * @deprecated
+     */
     getColumnCount: function() {
         return this.deprecated('getColumnCount()', 'getActiveColumnCount()', '1.0.6', arguments);
     },
@@ -1298,44 +1309,33 @@ var Behavior = Base.extend('Behavior', {
     },
 
     /**
-     * @summary Get the given data controller.
-     * @param {string} type
-     * @returns {undefined|*} The data controller; or `undefined` if data controller unknown to data model.
      * @memberOf Behavior#
+     * @deprecated
      */
     getController: function(type) {
-        return this.dataModel.getController(type);
+        return this.deprecated('getController()', '', '1.4.0', arguments, 'No Longer Supported');
     },
 
     /**
-     * @summary Set the given data controller(s).
-     * @desc Triggers a shape change.
-     * @param {string} typeOrHashOfTypes - One of:
-     * * **object** - Hash of multiple data controllers, by type.
-     * * **string** - Type of the single data controller given in `controller`.
-     * @param {dataControlInterface} [controller] - Only required when 'hash' is a string; omit when `hash` is an object.
-     * @returns {object} - Hash of all results, by type. Each member will be:
-     * * The given data controller for that type when defined.
-     * * A new "null" data controller, generated by the data model when the given data controller for that type was `undefined`.
-     * * `undefined` - The data controller was unknown to the data model..
      * @memberOf Behavior#
+     * @deprecated
      */
     setController: function(typeOrHashOfTypes, controller) {
-        var results = this.dataModel.setController(typeOrHashOfTypes, controller);
-        this.changed();
-        return results;
+        return this.deprecated('setController()', '', '1.4.0', arguments, 'No Longer Supported');
     },
-
+    /**
+     * @memberOf Behavior#
+     * @deprecated
+     */
     prop: function(type, columnIndex, keyOrHash, value) {
-        var result = this.dataModel.prop.apply(this.dataModel, arguments);
-        if (result === undefined) {
-            this.changed();
-        }
-        return result;
+        return this.deprecated('prop()', '', '1.4.0', arguments, 'No Longer Supported');
     },
-
+    /**
+     * @memberOf Behavior#
+     * @deprecated
+     */
     get charMap() {
-        return this.dataModel.charMap;
+        return this.deprecated('charMap()', '', '1.4.0', arguments, 'No Longer Supported');
     },
 
     getColumns: function() {
@@ -1377,23 +1377,28 @@ var Behavior = Base.extend('Behavior', {
         return this.dataModel.getData();
     },
 
+    /**
+     * @memberOf Behavior#
+     * @deprecated
+     */
     getFilteredData: function() {
-        return this.deprecated('getIndexedData()', 'getIndexedData', '1.2.0', arguments);
+        return this.deprecated('getIndexedData()', '', '1.4.0', arguments, 'No Longer Supported');
     },
+    /**
+     * @memberOf Behavior#
+     * @deprecated
+     */
     getIndexedData: function() {
-       this.dataModel.getIndexedData();
+        return this.deprecated('getIndexedData()', '', '1.4.0', arguments, 'No Longer Supported');
+    },
+    /**
+     * @memberOf Behavior#
+     * @deprecated
+     */
+    applyAnalytics: function() {
+        return this.deprecated('applyAnalytics()', '', '1.4.0', arguments, 'No Longer Supported');
     }
 });
-
-// synonyms
-
-/**
- * Synonym of {@link Behavior#reindex}.
- * @name applyAnalytics
- * @deprecated
- * @memberOf Behavior#
- */
-Behavior.prototype.applyAnalytics = Behavior.prototype.reindex;
 
 
 // mix-ins
