+++ src/behaviors/JSON.js
@@ -23,9 +23,6 @@ var JSON = Behavior.extend('behaviors.JSON', {
      */
     initialize: function(grid, options) {
         this.setData(options);
-        if (options.pipeline) {
-            this.setPipeline(options.pipeline);
-        }
     },
 
     features: [
@@ -110,43 +107,19 @@ var JSON = Behavior.extend('behaviors.JSON', {
     },
 
     /**
-     * @see {@link dataModels.JSON#setPipeline}
-     * @param {object} [DataSources] - New pipeline description. _(See {@link dataModels.JSON#setPipeline}.)_
-     * @param {object} [options] - Takes first argument position when `DataSources` omitted. _(See {@link dataModels.JSON#setPipeline}.)_
-     * @param {boolean} [options.apply=true] Apply data transformations to the new data.
+     * @deprecated
      * @memberOf behaviors.JSON.prototype
      */
     setPipeline: function(DataSources, options) {
-        this.dataModel.setPipeline.apply(this.dataModel, arguments);
-
-        if (!Array.isArray(DataSources)) {
-            options = DataSources;
-        }
-
-        if (!options || options.apply === undefined || options.apply) {
-            this.reindex();
-        }
+        return this.deprecated('setPipeline()', '', '1.4.0', arguments, 'No Longer Supported');
     },
 
     /**
-     * Pop pipeline stack.
-     * @see {@link dataModels.JSON#unstashPipeline}
-     * @param {string} [whichStash]
-     * @param {object} [options] - Takes first argument position when `DataSources` omitted.
-     * @param {boolean} [options.apply=true] Apply data transformations to the new data.
+     * @deprecated
      * @memberOf behaviors.JSON.prototype
      */
     unstashPipeline: function(stash, options) {
-        if (typeof stash === 'object') {
-            options = stash;
-            stash = undefined;
-        }
-
-        this.dataModel.unstashPipeline(stash);
-
-        if (!options || options.apply === undefined || options.apply) {
-            this.reindex();
-        }
+        return this.deprecated('unstashPipeline()', '', '1.4.0', arguments, 'No Longer Supported');
     },
 
     /**
@@ -160,7 +133,6 @@ var JSON = Behavior.extend('behaviors.JSON', {
      * Passed as 2nd param to {@link dataModel.JSON#setData}.
      * @param {function|object} [options.calculators] - Array of calculators or function returning same.
      * Passed as 3rd param to {@link dataModel.JSON#setData}.
-     * @param {boolean} [options.apply=true] Apply data transformations to the new data.
      */
     setData: function(dataRows, options) {
         if (!(Array.isArray(dataRows) || typeof dataRows === 'function')) {
@@ -182,8 +154,7 @@ var JSON = Behavior.extend('behaviors.JSON', {
 
         var grid = this.grid,
             schema = this.unwrap(options.schema), // *always* define a new schema on reset
-            schemaChanged = schema || !this.subgrids.lookup.data.schema.length, // schema will change if a new schema was provided OR data model has an empty schema now, which triggers schema generation on setData below
-            reindex = options.apply === undefined || options.apply; // defaults to true
+            schemaChanged = schema || !this.subgrids.lookup.data.schema.length; // schema will change if a new schema was provided OR data model has an empty schema now, which triggers schema generation on setData below
 
         // Inform interested data models of data.
         this.subgrids.forEach(function(dataModel) {
@@ -194,10 +165,6 @@ var JSON = Behavior.extend('behaviors.JSON', {
 
         if (grid.cellEditor) {
             grid.cellEditor.cancelEditing();
-        }
-
-        if (reindex) {
-            this.reindex();
         }
 
         if (schemaChanged) {
@@ -254,7 +221,7 @@ var JSON = Behavior.extend('behaviors.JSON', {
 
     getSelections: function() {
         return this.grid.selectionModel.getSelections();
-    },
+    }
 });
 
 module.exports = JSON;
