+++ src/lib/index.js
@@ -2,7 +2,6 @@
 
 module.exports = {
     cellEventFactory: require('./cellEventFactory'),
-    DataSourceOrigin: require('./DataSourceOrigin'),
     dynamicProperties: require('./dynamicProperties'),
     graphics: require('./graphics'),
     Canvas: require('./Canvas'),
